requirements.txt
google-generativeai
pillow
requests
python-dotenv